+++ example/integration_test/downloader_integration_test.dart
@@ -50,6 +50,10 @@ const uploadBinaryTestUrl =
     'https://avmaps-dot-bbflightserver-hrd.appspot.com/public/test_upload_binary_file';
 const uploadMultiTestUrl =
     'https://avmaps-dot-bbflightserver-hrd.appspot.com/public/test_multi_upload_file';
+const dataTaskGetUrl = 'https://httpbin.org/get';
+const dataTaskPostUrl = 'https://httpbin.org/post';
+final dataTaskHeaders = {'accept': 'application/json'};
+
 const urlWithContentLengthFileSize = 6207471;
 
 const defaultFilename = 'google.html';
@@ -124,6 +128,7 @@ void main() {
     uploadTask = UploadTask(url: uploadTestUrl, filename: uploadFilename);
     uploadTaskBinary =
         uploadTask.copyWith(url: uploadBinaryTestUrl, post: 'binary');
+
     // copy the test files to upload from assets to documents directory
     Directory directory = await getApplicationDocumentsDirectory();
     for (final filename in [uploadFilename, uploadFilename2]) {
@@ -2313,7 +2318,7 @@ void main() {
     //   expect(await (File(await task.filePath())).length(), equals(59673498));
     //   expect(statusCallbackCounter, greaterThanOrEqualTo(9)); // min 2 pause
     // });
-//TODO put back multipause and resume
+//TODO put back multi pause and resume
     testWidgets('Pause and resume a convenience download',
         (widgetTester) async {
       task = DownloadTask(
@@ -2394,8 +2399,7 @@ void main() {
         lastStatus = TaskStatus.enqueued; // must change to paused
         final oldLastProgress = lastProgress;
         lastProgress = -1; // must change to a real value
-        await downloader
-            .retrieveLocallyStoredData(); // triggers status & prog update
+        await downloader.retrieveLocallyStoredData(); // triggers updates
         downloader.retrievedLocallyStoredData = false; // reset for testing
         await Future.delayed(const Duration(milliseconds: 500));
         expect(lastStatus, equals(TaskStatus.paused));
@@ -2543,7 +2547,7 @@ void main() {
         success = await FileDownloader().openFile(task: task);
         expect(success, isFalse);
       } else {
-        expect(success, isFalse); // on Android cannot access docsdir
+        expect(success, isFalse); // on Android cannot access docs dir
         // change to a .txt file
         final filePath = await task.filePath();
         await File(filePath).rename(join(
@@ -3294,6 +3298,166 @@ void main() {
           expect(filename, equals(task.filename));
         }
       }
+    });
+  });
+
+  group('DataTask', () {
+    test('dataTask get', () async {
+      var lastUpdate = TaskStatusUpdate(task, TaskStatus.paused);
+      FileDownloader().registerCallbacks(taskStatusCallback: (update) {
+        lastUpdate = update;
+      });
+      final t = DataTask(url: dataTaskGetUrl, headers: dataTaskHeaders);
+      expect(await FileDownloader().enqueue(t), isTrue);
+      await Future.delayed(const Duration(seconds: 1));
+      expect(lastUpdate.status, equals(TaskStatus.complete));
+      final json = jsonDecode(lastUpdate.responseBody!);
+      final args = json['args'] as Map<String, dynamic>;
+      expect(args.isEmpty, isTrue);
+      expect(json['headers']['Accept'], equals('application/json'));
+    });
+
+    test('dataTask post no data', () async {
+      var lastUpdate = TaskStatusUpdate(task, TaskStatus.paused);
+      FileDownloader().registerCallbacks(taskStatusCallback: (update) {
+        lastUpdate = update;
+      });
+      final t = DataTask(
+          url: dataTaskPostUrl,
+          headers: dataTaskHeaders); // without 'POST' will fail
+      expect(await FileDownloader().enqueue(t), isTrue);
+      await Future.delayed(const Duration(seconds: 1));
+      expect(lastUpdate.status, equals(TaskStatus.failed));
+      final t2 = DataTask(
+          url: dataTaskPostUrl,
+          headers: dataTaskHeaders,
+          httpRequestMethod: 'POST'); // with 'POST' will succeed
+      expect(await FileDownloader().enqueue(t2), isTrue);
+      await Future.delayed(const Duration(seconds: 1));
+      expect(lastUpdate.status, equals(TaskStatus.complete));
+      print(lastUpdate.responseBody);
+      final json = jsonDecode(lastUpdate.responseBody!);
+      final args = json['args'] as Map<String, dynamic>;
+      expect(args.isEmpty, isTrue);
+      expect(json['headers']['Accept'], equals('application/json'));
+      expect((json['data'] as String).isEmpty, isTrue);
+    });
+
+    test('dataTask post with data', () async {
+      var lastUpdate = TaskStatusUpdate(task, TaskStatus.paused);
+      FileDownloader().registerCallbacks(taskStatusCallback: (update) {
+        lastUpdate = update;
+      });
+      final t = DataTask(
+          url: dataTaskPostUrl,
+          headers: dataTaskHeaders,
+          post: 'My data'); // should auto set 'POST' method
+      expect(await FileDownloader().enqueue(t), isTrue);
+      await Future.delayed(const Duration(seconds: 1));
+      print(lastUpdate.responseStatusCode);
+      expect(lastUpdate.status, equals(TaskStatus.complete));
+      print(lastUpdate.responseBody);
+      final json = jsonDecode(lastUpdate.responseBody!);
+      final args = json['args'] as Map<String, dynamic>;
+      expect(args.isEmpty, isTrue);
+      expect(json['headers']['Accept'], equals('application/json'));
+      expect((json['data'] as String), equals('My data'));
+    });
+
+    test('dataTask post with json data', () async {
+      var lastUpdate = TaskStatusUpdate(task, TaskStatus.paused);
+      FileDownloader().registerCallbacks(taskStatusCallback: (update) {
+        lastUpdate = update;
+      });
+      var jsonData = {'key': 'value'};
+      final t = DataTask(
+          url: dataTaskPostUrl, headers: dataTaskHeaders, json: jsonData);
+      expect(await FileDownloader().enqueue(t), isTrue);
+      await Future.delayed(const Duration(seconds: 1));
+      print(lastUpdate.responseStatusCode);
+      expect(lastUpdate.status, equals(TaskStatus.complete));
+      print(lastUpdate.responseBody);
+      final json = jsonDecode(lastUpdate.responseBody!);
+      final args = json['args'] as Map<String, dynamic>;
+      expect(args.isEmpty, isTrue);
+      expect(json['headers']['Accept'], equals('application/json'));
+      expect((json['data'] as String), equals("{\"key\":\"value\"}"));
+      expect((json['json'] as Map<String, dynamic>), equals(jsonData));
+    });
+
+    test('dataTask with error', () async {
+      var lastUpdate = TaskStatusUpdate(task, TaskStatus.paused);
+      FileDownloader().registerCallbacks(taskStatusCallback: (update) {
+        lastUpdate = update;
+      });
+      final t = DataTask(
+          url: 'https://httpbin.org/status/400', // force 400 code
+          headers: dataTaskHeaders);
+      expect(await FileDownloader().enqueue(t), isTrue);
+      await Future.delayed(const Duration(seconds: 1));
+      expect(lastUpdate.status, equals(TaskStatus.failed));
+      final exception = lastUpdate.exception!;
+      print(exception);
+      expect(exception is TaskHttpException, isTrue);
+      expect(exception.description.toUpperCase(), equals('BAD REQUEST'));
+      expect((exception as TaskHttpException).httpResponseCode, equals(400));
+    });
+
+    test('dataTask with retries', () async {
+      var lastUpdate = TaskStatusUpdate(task, TaskStatus.paused);
+      var retryCount = 0;
+      FileDownloader().registerCallbacks(taskStatusCallback: (update) {
+        lastUpdate = update;
+        if (update.status == TaskStatus.waitingToRetry) {
+          retryCount++;
+        }
+      });
+      final t = DataTask(
+          url: 'https://httpbin.org/status/400', // force 400 code
+          headers: dataTaskHeaders,
+          retries: 2);
+      expect(await FileDownloader().enqueue(t), isTrue);
+      await Future.delayed(const Duration(seconds: 10));
+      expect(lastUpdate.status, equals(TaskStatus.failed));
+      expect(lastUpdate.task.retries, equals(2));
+      expect(lastUpdate.task.retriesRemaining, equals(0));
+      expect(retryCount, equals(2));
+    });
+
+    test('cancel dataTask', () async {
+      /// Cancellation is only relevant when the task is waiting to retry.
+      /// Cancellation of a running DataTask has no effect
+      var lastUpdate = TaskStatusUpdate(task, TaskStatus.paused);
+      FileDownloader().registerCallbacks(taskStatusCallback: (update) {
+        lastUpdate = update;
+        if (update.status == TaskStatus.waitingToRetry) {
+          // cancel with some delay (otherwise the waitingToRetry status has
+          // not been registered yet)
+          Future.delayed(
+              const Duration(milliseconds: 500),
+              () => FileDownloader()
+                  .cancelTaskWithId(update.task.taskId)
+                  .then((success) => expect(success, isTrue)));
+        }
+      });
+      final t = DataTask(
+          url: 'https://httpbin.org/status/400', // force 400 code
+          headers: dataTaskHeaders,
+          retries: 2);
+      expect(await FileDownloader().enqueue(t), isTrue);
+      await Future.delayed(const Duration(seconds: 3));
+      expect(lastUpdate.status, equals(TaskStatus.canceled));
+      expect(lastUpdate.task.retriesRemaining, greaterThan(0));
+    });
+
+    test('transmit (wait for completion)', () async {
+      final t = DataTask(url: dataTaskGetUrl, headers: dataTaskHeaders);
+      final result = await FileDownloader().transmit(t);
+      expect(result.status, equals(TaskStatus.complete));
+      final json = jsonDecode(result.responseBody!);
+      final args = json['args'] as Map<String, dynamic>;
+      expect(args.isEmpty, isTrue);
+      expect(json['headers']['Accept'], equals('application/json'));
     });
   });
 }