+++ ios/Classes/BDPlugin.swift
@@ -214,7 +214,7 @@ public class BDPlugin: NSObject, FlutterPlugin, UNUserNotificationCenterDelegate
             // ParallelDownloadTask itself is not part of a urlSession, so handled separately
             baseRequest.httpMethod = "HEAD" // override
             return await scheduleParallelDownload(task: task, taskDescription: taskDescription, baseRequest: baseRequest, resumeData: resumeDataAsBase64String)
-        } else if isDownloadTask(task: task)
+        } else if isDownloadTask(task: task) || isDataTask(task: task)
         {
             return scheduleDownload(task: task, taskDescription: taskDescription, baseRequest: baseRequest, resumeData: resumeData)
         } else
@@ -268,6 +268,7 @@ public class BDPlugin: NSObject, FlutterPlugin, UNUserNotificationCenterDelegate
                 return false
             }
             // binary post can use uploadTask fromFile method
+            request.setValue(task.mimeType, forHTTPHeaderField: "Content-Type")
             request.setValue("attachment; filename=\"\(task.filename)\"", forHTTPHeaderField: "Content-Disposition")
             let urlSessionUploadTask = UrlSessionDelegate.urlSession!.uploadTask(with: request, fromFile: filePath)
             urlSessionUploadTask.taskDescription = taskDescription