+++ lib/src/desktop/isolate.dart
@@ -13,6 +13,7 @@ import 'package:logging/logging.dart';
 
 import '../models.dart';
 import '../task.dart';
+import 'data_isolate.dart';
 import 'desktop_downloader.dart';
 import 'download_isolate.dart';
 import 'parallel_download_isolate.dart';
@@ -95,7 +96,8 @@ Future<void> doTask((RootIsolateToken, SendPort) isolateArguments) async {
           sendPort),
       DownloadTask() => doDownloadTask(task, filePath, resumeData, isResume,
           requestTimeout ?? const Duration(seconds: 60), sendPort),
-      UploadTask() => doUploadTask(task, filePath, sendPort)
+      UploadTask() => doUploadTask(task, filePath, sendPort),
+      DataTask() => doDataTask(task, sendPort)
     };
   }
   receivePort.close();